requirements.txt
streamlit
PyPDF2
torch
transformers
chromadb
nltk
numpy
pandas
google-generativeai
sentence-transformers
python-magic
python-magic-bin; platform_system == "Windows"
